Add company configuration module with API and QR code generation
- Introduced CompanyConfig entity, service, and controller for managing company configuration.
- Implemented endpoints for retrieving and updating company configuration.
- Added validation rules for company configuration fields.
- Created NcfSequence entity and repository for managing NCF sequences.
- Integrated QR code generation for invoices using ZXing library.
- Updated README with documentation for the new company configuration module and API endpoints.

+++ src/main/java/com/happydev/prestockbackend/dto/SaleItemDto.java
@@ -6,6 +6,7 @@ import jakarta.validation.constraints.Positive;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 
 @Getter @Setter
 public class SaleItemDto {
@@ -20,5 +21,5 @@ public class SaleItemDto {
 
     @NotNull(message = "Price cannot be null")
     @Positive(message = "Price must be positive")
-    private double unitPrice; // Precio unitario al momento de la venta
+    private BigDecimal unitPrice;
 }
